Extract map loading overlay into hook, component, and shared type
