+++ frontend/src/types/map.ts
@@ -1,0 +1,1 @@
+export type MapStatus = "idle" | "loading" | "error" | "ready";
